show profile

+++ src/database/queries/post.js
@@ -18,14 +18,14 @@ const getUserPosts = (id) => {
 };
 const getAllPosts = () => {
   const sql = {
-    text: 'SELECT posts.title ,posts.id,posts.details,posts.created_at,posts.image_url,users.username,json_agg(json_build_object(\'comment_id\',comments.user_id,\'comment\',comments.comments,\'commenter\',users_comment.username,\'creted_at\',comments.created_at)) AS comments,SUM(CASE WHEN votes.vote=\'upvote\' THEN 1 WHEN votes.vote=\'downvote\' THEN -1 WHEN  votes.vote=\'none\' THEN 0 ELSE 0 END) AS up_votes FROM posts LEFT JOIN users ON posts.user_id =users.id LEFT JOIN votes ON posts.id =votes.post_id LEFT JOIN comments ON posts.id =comments.post_id LEFT JOIN users AS users_comment ON comments.user_id=users_comment.id GROUP BY posts.id,users.username ORDER BY up_votes DESC;',
+    text: 'SELECT posts.title ,posts.id,posts.details,posts.created_at,posts.image_url,users.username,json_agg(json_build_object(\'comment_id\',comments.user_id,\'comments\',comments.comments,\'username\',users_comment.username,\'created_at\',comments.created_at)) AS comments,SUM(CASE WHEN votes.vote=\'upvote\' THEN 1 WHEN votes.vote=\'downvote\' THEN -1 WHEN  votes.vote=\'none\' THEN 0 ELSE 0 END) AS up_votes FROM posts LEFT JOIN users ON posts.user_id =users.id LEFT JOIN votes ON posts.id =votes.post_id LEFT JOIN comments ON posts.id =comments.post_id LEFT JOIN users AS users_comment ON comments.user_id=users_comment.id GROUP BY posts.id,users.username ORDER BY up_votes DESC;',
   };
   return connection.query(sql);
 };
-const deletePost = (id) => {
+const deletePost = (postId, user_id) => {
   const sql = {
-    text: 'DELETE FROM posts WHERE posts.id=$1 RETURNING id',
-    values: [id],
+    text: 'DELETE FROM posts WHERE posts.id=$1 AND posts.user_id =$2 RETURNING id',
+    values: [postId, user_id],
   };
   return connection.query(sql);
 };
